@@ -1,4 +1,4 @@
-class CreatePosts < ActiveRecord::Migration
+class CreatePosts < ActiveRecord::Migration[5.0]
   def change
     create_table :posts do |t|
       t.string :title
@@ -6,7 +6,7 @@ class CreatePosts < ActiveRecord::Migration
       t.text :description
       t.string :slug
 
-      t.timestamps null: false
+      t.timestamps
     end
     add_index :posts, :slug, unique: true
   end
