@@ -1,49 +1,48 @@
 require 'test_helper'
 
-class PostsControllerTest < ActionController::TestCase
+class PostsControllerTest < ActionDispatch::IntegrationTest
   setup do
     @post = posts(:one)
   end
 
   test "should get index" do
-    get :index
+    get posts_url
     assert_response :success
-    assert_not_nil assigns(:posts)
   end
 
   test "should get new" do
-    get :new
+    get new_post_url
     assert_response :success
   end
 
   test "should create post" do
     assert_difference('Post.count') do
-      post :create, post: { body: @post.body, description: @post.description, slug: @post.slug, title: @post.title }
+      post posts_url, params: { post: { body: @post.body, description: @post.description, slug: @post.slug, title: @post.title } }
     end
 
-    assert_redirected_to post_path(assigns(:post))
+    assert_redirected_to post_url(Post.last)
   end
 
   test "should show post" do
-    get :show, id: @post
+    get post_url(@post)
     assert_response :success
   end
 
   test "should get edit" do
-    get :edit, id: @post
+    get edit_post_url(@post)
     assert_response :success
   end
 
   test "should update post" do
-    patch :update, id: @post, post: { body: @post.body, description: @post.description, slug: @post.slug, title: @post.title }
-    assert_redirected_to post_path(assigns(:post))
+    patch post_url(@post), params: { post: { body: @post.body, description: @post.description, slug: @post.slug, title: @post.title } }
+    assert_redirected_to post_url(@post)
   end
 
   test "should destroy post" do
     assert_difference('Post.count', -1) do
-      delete :destroy, id: @post
+      delete post_url(@post)
     end
 
-    assert_redirected_to posts_path
+    assert_redirected_to posts_url
   end
 end
